Fix the issue of medical document cannot be deleted.
#8801 - As a Developer, I will have fixed issues in MHC in order to make the system more stable and reliable

+++ pcm/src/main/java/gov/samhsa/mhc/pcm/service/clinicaldata/ClinicalDocumentServiceImpl.java
@@ -168,8 +168,13 @@ public class ClinicalDocumentServiceImpl implements ClinicalDocumentService {
      */
     @Override
     public void deleteClinicalDocument(ClinicalDocumentDto clinicalDocumentDto) {
-        ClinicalDocument clinicalDocument = getClinicalDocumenFromDto(clinicalDocumentDto);
-        clinicalDocumentRepository.delete(clinicalDocument);
+        ClinicalDocument clinicalDocument = clinicalDocumentRepository.findClinicalDocumentById(Long.parseLong(clinicalDocumentDto.getId()));
+        if(clinicalDocument != null){
+            clinicalDocumentRepository.delete(clinicalDocument);
+        }else {
+            logger.error("Cannot get document to be deleted by id.");
+        }
+
     }
 
     /*
